+++ changelog.md
@@ -1,3 +1,9 @@
+# Version 0.2.0
+
+- Better support for generics with long syntax
+    - `#[static_dispatch::dispatch(impl<'a> SomethingBehavior for Something<'a>)]`
+
+
 # Version 0.1.0
 
 - Initial version